Fixed commands, messagebox, mod desc, other small fixes
The mod now compiles and runs! Lootbox changes
Attempted and failed to fix fireballs, and raytracing stuff is still completely broken

+++ src/main/java/io/github/icrazyblaze/twitchmod/util/timers/TimerSystem.java
@@ -49,13 +49,13 @@ public class TimerSystem {
         }
 
         // Death timer
-        if (deathTimerSeconds > 0) {
-            deathTimerSeconds--;
-        } else if (deathTimerSeconds == 0) {
-
-            PlayerHelper.player().kill();
-            deathTimerEnabled = false;
-
+        if (deathTimerEnabled) {
+            if (deathTimerSeconds > 0) {
+                deathTimerSeconds--;
+            } else if (deathTimerSeconds == 0) {
+                PlayerHelper.player().kill();
+                deathTimerEnabled = false;
+            }
         }
 
         // Frenzy mode timer
